[#150] Type instantiation redesign

+++ test/language/packages/java/DefaultPackageTest.java
@@ -1,5 +1,7 @@
 import static org.junit.Assert.assertEquals;
 
+import java.math.BigInteger;
+
 import org.junit.Test;
 
 public class DefaultPackageTest
@@ -9,10 +11,10 @@ public class DefaultPackageTest
     {
         // just test that DefaultPackageStructure is available in default package
         DefaultPackageStructure structure = new DefaultPackageStructure((short)4);
-        structure.setValue(10);
+        structure.setValue(BigInteger.valueOf(10));
         structure.setTopStructure(new default_package_import.top.TopStructure((short)1, 1234));
         structure.setChildStructure(new Child(0xdeadbeef));
-        assertEquals(10, structure.getValue());
+        assertEquals(BigInteger.valueOf(10), structure.getValue());
         assertEquals(1, structure.getTopStructure().getType());
         assertEquals(1234, structure.getTopStructure().getData());
         assertEquals(0xdeadbeef, structure.getChildStructure().getValue());
